Creation of rocketchat:ui-utils to help decouple rocketchat:lib and rocketchat:authz
