+++ .meteor/packages
@@ -207,3 +207,4 @@ rocketchat:metrics
 rocketchat:callbacks
 rocketchat:notifications
 rocketchat:promises
+rocketchat:ui-utils
