+++ .meteor/versions
@@ -254,6 +254,7 @@ rocketchat:ui-login@0.1.0
 rocketchat:ui-master@0.1.0
 rocketchat:ui-message@0.1.0
 rocketchat:ui-sidenav@0.1.0
+rocketchat:ui-utils@0.0.1
 rocketchat:ui-vrecord@0.0.1
 rocketchat:user-data-download@1.0.0
 rocketchat:utils@0.0.1
