+++ packages/rocketchat-lib/package.js
@@ -31,6 +31,7 @@ Package.onUse(function(api) {
 	api.use('rocketchat:callbacks');
 	api.use('rocketchat:notifications');
 	api.use('rocketchat:promises');
+	api.use('rocketchat:ui-utils');
 	api.use('rocketchat:accounts');
 	api.use('modules');
 	api.use('rocketchat:i18n');
