+++ packages/rocketchat-ui-utils/client/index.js
@@ -1,0 +1,7 @@
+import { AdminBox } from './lib/AdminBox';
+import { modal } from './lib/modal';
+
+export {
+	AdminBox,
+	modal,
+};
